pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azuraforge-api"
version = "0.1.2" 
description = "The API server for the AzuraForge Platform."
requires-python = ">=3.10"

dependencies = [
    "azuraforge-learner @ git+https://github.com/AzuraForge/learner.git@main",
    "azuraforge-applications @ git+https://github.com/AzuraForge/applications.git@main",
    "azuraforge-dbmodels @ git+https://github.com/AzuraForge/dbmodels.git@main",
    "celery[redis]",
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-dotenv",
    "pyyaml",
    "redis",
    "python-semantic-release",
    "passlib[bcrypt]", # <-- YENİ: Parola hashleme için
    "python-jose[cryptography]" # <-- YENİ: JWT işlemleri için
]

[project.scripts]
start-api = "azuraforge_api.main:run_server"

[project.urls]
"Homepage" = "https://github.com/AzuraForge/api"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.semantic_release]
branch = "main"
version_variable = "pyproject.toml:project.version"
upload_to_pypi = false
major_on_zero = false
changelog_file = "CHANGELOG.md"